Crate ralloc_shim

Symbols and externs that ralloc depends on.

This crate provides implementation/import of these in Linux, BSD, and Mac OS.

Important

You CANNOT use libc library calls, due to no guarantees being made about allocations of the functions in the POSIX specification. Therefore, we use the system calls directly.

Modules

config
Configuration.

debug
Bindings to debuggers.

syscalls
System calls.

thread_destructor
Thread destructors.