Module ralloc_shim::config

Configuration.

This module contains anything which can be tweaked and customized to the users preferences.

Constants

FRAGMENTATION_SCALE
The fragmentation scale constant.

LOCAL_MEMTRIM_LIMIT
The local memtrim limit.

LOCAL_MEMTRIM_STOP
The local memtrim chock.

MIN_LOG_LEVEL
The minimum log level.

OS_MEMTRIM_LIMIT
The memtrim limit.

OS_MEMTRIM_WORTHY
Minimum size before a block is worthy to memtrim.

Functions

default_oom_handler
The default OOM handler.

extra_brk
Canonicalize a BRK request.

extra_fresh
Canonicalize a fresh allocation.

log
Write to the log.